+++ tests/test_query.py
@@ -33,7 +33,7 @@ class TestRequest:
         no.session.session.cookies.clear()
         r = no.session.method("get", "pages", id_="878d628488d94894ab14f9b872cd6870")
         assert isinstance(r, dict)
-        assert len(r) == 12
+        assert len(r) == 13
         assert r["object"] == "page"
         assert r["archived"] is False
         assert r["id"] == "878d6284-88d9-4894-ab14-f9b872cd6870"
@@ -41,7 +41,7 @@ class TestRequest:
     def test_method__patch_empty(self, no):
         r = no.session.method("patch", "pages", id_="878d628488d94894ab14f9b872cd6870")
         assert isinstance(r, dict)
-        assert len(r) == 12
+        assert len(r) == 13
         assert r["object"] == "page"
         assert r["archived"] is False
         assert r["id"] == "878d6284-88d9-4894-ab14-f9b872cd6870"
@@ -145,7 +145,7 @@ class TestRequest:
     def test_method__limit_patch(self, no):  # limit in patch mode is ignored
         r = no.session.method("patch", "pages", id_="878d628488d94894ab14f9b872cd6870", limit=2)
         assert isinstance(r, dict)
-        assert len(r) == 12
+        assert len(r) == 13
         assert r["object"] == "page"
         assert r["archived"] is False
         assert r["id"] == "878d6284-88d9-4894-ab14-f9b872cd6870"